Authorize UserController

+++ app/Http/Controllers/UserController.php
@@ -4,12 +4,15 @@ namespace App\Http\Controllers;
 
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Hash;
 
 class UserController extends Controller
 {
     public function index()
     {
+        Gate::authorize('users');
+
         $users = User::query();
 
         if (request('role')) {
@@ -23,11 +26,15 @@ class UserController extends Controller
 
     public function create()
     {
+        Gate::authorize('users');
+
         return view('users.create');
     }
 
     public function store(Request $request)
     {
+        Gate::authorize('users');
+
         $data = $request->validate([
             'name' => ['required', 'string', 'max:255'],
             'email' => ['required', 'email', 'max:255'],
@@ -55,11 +62,15 @@ class UserController extends Controller
 
     public function edit(User $user)
     {
+        Gate::authorize('users');
+
         return view('users.edit', compact('user'));
     }
 
     public function update(Request $request, User $user)
     {
+        Gate::authorize('users');
+
         $data = $request->validate([
             'name' => ['required', 'string', 'max:255'],
             'email' => ['required', 'email', 'max:255'],
@@ -87,6 +98,8 @@ class UserController extends Controller
 
     public function destroy(User $user)
     {
+        Gate::authorize('users');
+
         $user->delete();
 
         alert()->success('حذف موفق', 'حذف کاربر با موفقیت انجام شد');
@@ -96,12 +109,16 @@ class UserController extends Controller
 
     public function deleted()
     {
+        Gate::authorize('users');
+
         $users = User::onlyTrashed()->latest()->paginate(20);
         return view('users.deleted', compact('users'));
     }
 
     public function restore($id)
     {
+        Gate::authorize('users');
+
         $user = User::where('id', $id)->withTrashed()->first();
 
         $user->restore();
@@ -113,6 +130,8 @@ class UserController extends Controller
 
     public function forceDelete($id)
     {
+        Gate::authorize('users');
+
         $user = User::where('id', $id)->withTrashed()->first();
 
         $user->forceDelete();
